+++ version.log
@@ -1,143 +1,18 @@
 
-fithic_1.0.0.tgz - 01/02/2014
+fithic_2.0.0.tgz - 06/12/2018
 -------------------------------
-Initial release.
+Initial release of Fit-Hi-C version 2.0
 __________________________________________________________________
 
-fithic_1.0.1.tgz - 01/19/2014
--------------------------------
-Bug fixes/changes:
-1. Output file names are changed to be consistent with README and 
-	header lines are added to each output file.
-2. Power-law fit part is removed from the implementation.
-3. --nobinning option is removed. Now default is using 100 bins.
-4. Fixed the warning from matplotlib.use in --visual mode.
+Notable changes:
+* Significantly faster and uses less memory
+* Required '--resolution, -r' argument
+* Required '--outDir, -o' argument
+* Specific to Python3
+* Recommended bias file use
+* Change in stdout outputs
+* Creation of fithic.log file detailing previous values
+* Complete revamp of utils/ with additional scripts provided
+* Default number of spline passes is 1. Use -p 2 for 2 passes etc.
 
-__________________________________________________________________
 
-fithic_1.0.3.tgz - 08/01/2017
--------------------------------
-Changes:
-1. Updated directory structure to be configured for pip
-2. Removed R dependencies
-3. Fixed Python2.7 version requirement
-
-__________________________________________________________________
-
-fithic_1.0.4.tgz - 08/09/2017
--------------------------------
-Changes:
-1. Updated structure of setup.py to allow tests
-2. Included changes to allow conda installation
-__________________________________________________________________
-
-fithic_1.0.5.tgz - 08/09/2017
--------------------------------
-Changes:
-1. Modified location of tests directory 
-__________________________________________________________________
-
-fithic_1.0.6.tgz - 10/09/2017
--------------------------------
-Changes:
-1. Added new option '-x' this allows the user to specify 'intraOnly',
-    'interOnly', or 'allReg' as areas to analyze. Default option
-    is 'intraOnly'
-
-Bug Fixes:
-1. Updated error for Illegal Fragment pairs with better error message 
-2. Added NaN check 
-3. Made the output dir. if user has not already
-
-Sorry this took so long!
-__________________________________________________________________
-
-fithic_1.0.7.tgz - 10/26/2017
--------------------------------
-Bug Fixes:
-1. Fixed bug with -x option 
-2. Added conditionals to import matplotlib statement 
-
-__________________________________________________________________
-
-fithic_1.0.8.tgz - 12/11/2017 #Thank you to Ye Zheng for bringing
-                                this to our attention!
--------------------------------
-Bug Fixes:
-1. Isotonic regression is now strictly antitonic. 
-~~~During the spline fit phase of fithic, we are looking to smooth
-the contacts to a monotonically decreasing function. Previously,
-the IsotonicRegression function from sklearn was not making this
-assumption, and was empirically determining if the function
-was isotonic(increasing) or antitonic(decreasing). While it would
-typically determine it is antitonic and proceed normally, Ye found
-a case where the function determined the flip (isotonic) and this
-led to significant influences on fithic's 'significant contacts' 
-determination. 
-
-It should be obvious when this has occured, as the
-results would be significantly different; however, it is still 
-recommended all users update ASAP.
-__________________________________________________________________
-
-fithic_1.0.9.tgz - 12/19/2017 
--------------------------------
-Bug Fixes:
-1. Modified isotonic regression such that it's not version
-independent
-__________________________________________________________________
-
-fithic_1.1.0.tgz - 1/11/2018
--------------------------------
-Bug Fixes:
-1. Modified file parser to accept decimal's such that it works 
-with hicpro2fithic.py script
-__________________________________________________________________
-
-fithic_1.1.1.tgz - 1/11/2018
--------------------------------
-Bug Fixes:
-1. Thought of better fix to the problem addressed in 1.1.0
-__________________________________________________________________
-
-fithic_1.1.2.tgz - 02/05/2018
--------------------------------
-Bug Fixes:
-1. Fixed issue with optional import of matplotlib and pylab
-2. Made input compression agnostic (you don't have to gzip input
-files anymore! But we still highly recommend it!)
-
-Special thanks to aakashsur for all the ideas!
-__________________________________________________________________
-
-fithic_1.1.3.tgz - 02/20/2018
--------------------------------
-Bug Fixes:
-1. Fixed a persistent and vastly annoying bug with -x All 
-and -x intersOnly and p-values
-2. Makes default value of -l option 'fithic' so files are
-not hidden any longer
-3. Makes default value of -o option './FITHICOUTPUT' so 
-users don't lose their output to the abyss
-4. Attempted to fix matplotlib plotting issues per aakashsur's
-issues. Unclear if it will work for all matplotlib versions.
-Still an area of active work.
-
-Updates:
-1. README updated - it's pretty now! check it out!
-2. utils/ folder added to fithic contains two scripts
-    1. hicpro2fithic.py - a Python script to convert 
-    hicpro output to input for fithic. Also found in the
-    hicpro github
-    2. Create_Fragments_File.py - a Python script to 
-    automatically generate the fragsfile for Hi-C
-    data given a chromsome sizes file and a 
-    resolution for the data
-3. Added whole genome dataset to tests/ folder to test
-for interchromosomal interactions as well
-4. Added parasite dataset to tests/ folder to test for
-edge cases
-
-Special thanks to Vikash and Moshe for pointing out and helping 
-debug bug #1!
-__________________________________________________________________
